Add banner post type/taxo settings
Big restructuring to banner area. There are now settings to disable the banner area for specific post types and specific taxonomies. Banner image also has a mid-fallback now. If a single post or CPT doesn’t have a banner image uploaded, it will first check if the post type archive settings has an image, then fallback to the default image. This means we can have one image for the Blog and all single posts, and a differnt image for the Shop and all products, etc.